+++ src/HieDb/Query.hs
@@ -51,6 +51,11 @@ getExportsForModule :: HieDb -> ModuleName -> IO [ExportRow]
 getExportsForModule (getConn -> conn) mn =
   query conn "SELECT exports.* FROM exports JOIN mods USING (hieFile) WHERE mods.mod = ?" (Only mn)
 
+{-| Find all the modules that export an identifier |-}
+findExporters :: HieDb -> OccName -> IO [ModuleName]
+findExporters (getConn -> conn) occ =
+  query conn "SELECT mods.mod FROM exports JOIN mods USING (hieFile) WHERE occ = ?" (Only occ)
+
 {-| Lookup Unit associated with given ModuleName.
 HieDbErr is returned if no module with given name has been indexed
 or if ModuleName is ambiguous (i.e. there are multiple packages containing module with given name)