+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.4.0.0 -- 2021-06-29
 
-* Index module exports and add `getAllIndexedExports` and `getExportsForModule`
+* Index module exports
+* New queries `getAllIndexedExports`, `getExportsForModule`, and `findExporters`
 
 ## 0.3.0.1 -- 2021-01-27